move Bounty::Task and Bounty::Parser to their own files
